Support snapping, maximize/unmaximize, and catchall id
Snapping of clients is now supported, this will allow the snapping of
clients to either the left or right side of the monitor and at 50% width.

Maximizing and unmaximizing of clients can now be down too. The
maximizeclient function will now keep track of the previous client
position and geometry, only if that client was not
fullscreened/maximized.

The app_id "*" is supported in the Layout struct for applying rules to
all applications that are spawned.